refactor: Rework how we parse the option string.
 - Instead of checking the string for each option that exists, loop
   through the string to discover which options have been specified
 - Split it out into a parser that's arguably overkill for what we're
   currently doing, but consider this as a prefactor for https://github.com/google/keep-sorted/pull/29
