+++ keepsorted/keep_sorted_test.go
@@ -15,13 +15,10 @@
 package keepsorted
 
 import (
-	"fmt"
-	"reflect"
 	"strings"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 )
@@ -1526,183 +1523,5 @@ func TestLineGrouping(t *testing.T) {
 				t.Errorf("groupLines mismatch (-want +got):\n%s", diff)
 			}
 		})
-	}
-}
-
-func TestBlockOptions(t *testing.T) {
-	for _, tc := range []struct {
-		name           string
-		in             string
-		defaultOptions blockOptions
-
-		want    blockOptions
-		wantErr string
-	}{
-		{
-			name:           "DefaultOptions",
-			in:             "",
-			defaultOptions: defaultOptions,
-
-			want: defaultOptions,
-		},
-		{
-			name: "CommentMarker",
-			in:   "// keep-sorted-test start",
-
-			want: blockOptions{
-				commentMarker: "//",
-			},
-		},
-		{
-			name: "StickyComments",
-			in:   "// keep-sorted-test start sticky_comments=yes",
-
-			want: blockOptions{
-				StickyComments: true,
-				StickyPrefixes: map[string]bool{"//": true},
-				commentMarker:  "//",
-			},
-		},
-		{
-			name: "SimpleSwitch",
-			in:   "lint=yes",
-
-			want: blockOptions{Lint: true},
-		},
-		{
-			name: "SkipLines",
-			in:   "skip_lines=10",
-
-			want: blockOptions{SkipLines: 10},
-		},
-		{
-			name: "ErrorSkipLinesIsNegative",
-			in:   "skip_lines=-1",
-
-			wantErr: "skip_lines has invalid value: -1",
-		},
-		{
-			name: "ItemList",
-			in:   "prefix_order=a,b,c,d",
-
-			want: blockOptions{
-				PrefixOrder: []string{"a", "b", "c", "d"},
-			},
-		},
-		{
-			name: "ItemSet",
-			in:   "sticky_prefixes=a,b,c,d",
-
-			want: blockOptions{
-				StickyPrefixes: map[string]bool{
-					"a": true,
-					"b": true,
-					"c": true,
-					"d": true,
-				},
-			},
-		},
-		{
-			name: "ignore_prefixes",
-			in:   "ignore_prefixes=a,b,c,d",
-
-			want: blockOptions{
-				IgnorePrefixes: []string{"a", "b", "c", "d"},
-			},
-		},
-		{
-			name: "ignore_prefixes_ChecksLongestPrefixesFirst",
-			in:   "ignore_prefixes=DoSomething(,DoSomething({",
-
-			want: blockOptions{
-				IgnorePrefixes: []string{"DoSomething({", "DoSomething("},
-			},
-		},
-		{
-			name: "GroupPrefixesRequiresGrouping",
-			in:   "group_prefixes=a,b,c group=no",
-
-			wantErr: "group_prefixes may not be used with group=no",
-		},
-		{
-			name: "OptionInTrailingComment",
-			in:   "# keep-sorted-test start block=yes  # lint=yes",
-
-			want: blockOptions{
-				Block:         true,
-				Lint:          true,
-				commentMarker: "#",
-			},
-		},
-		{
-			name: "ErrorDoesNotStopParsing",
-			in:   "lint=nah case=no",
-			defaultOptions: blockOptions{
-				Lint:          true,
-				CaseSensitive: true,
-			},
-
-			want: blockOptions{
-				Lint:          true, // The default value should not change.
-				CaseSensitive: false,
-			},
-			wantErr: `option "lint" has unknown value "nah"`,
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			initZerolog(t)
-			got, err := parseBlockOptions(tc.in, tc.defaultOptions)
-			if err != nil {
-				if tc.wantErr == "" {
-					t.Errorf("parseBlockOptions(%q) = %v", tc.in, err)
-				} else if !strings.Contains(err.Error(), tc.wantErr) {
-					t.Errorf("parseBlockOptions(%q) = %v, expected to contain %q", tc.in, err, tc.wantErr)
-				}
-			}
-			if diff := cmp.Diff(tc.want, got, cmp.AllowUnexported(blockOptions{})); diff != "" {
-				t.Errorf("parseBlockOptions(%q) mismatch (-want +got):\n%s", tc.in, diff)
-			}
-
-			_ = got.String() // Make sure this doesn't panic.
-		})
-	}
-}
-
-func TestBlockOptions_ClonesDefaultOptions(t *testing.T) {
-	defaults := blockOptions{
-		StickyPrefixes: map[string]bool{},
-	}
-	_, err := parseBlockOptions("sticky_prefixes=//", defaults)
-	if err != nil {
-		t.Errorf("parseBlockOptions() = _, %v", err)
-	}
-	if diff := cmp.Diff(blockOptions{}, defaults, cmp.AllowUnexported(blockOptions{}), cmpopts.EquateEmpty()); diff != "" {
-		t.Errorf("defaults appear to have been modified (-want +got):\n%s", diff)
-	}
-}
-
-func TestBlockOptions_ClonesDefaultOptions_Reflection(t *testing.T) {
-	defaults := blockOptions{}
-	defaultOpts := reflect.ValueOf(&defaults).Elem()
-	var s []string
-	for i := 0; i < defaultOpts.NumField(); i++ {
-		val := defaultOpts.Field(i)
-		switch val.Kind() {
-		case reflect.Bool, reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64, reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Float32, reflect.Float64, reflect.Complex64, reflect.Complex128, reflect.String:
-			continue
-		case reflect.Slice:
-			val.Set(reflect.MakeSlice(val.Type(), 0, 0))
-			s = append(s, fmt.Sprintf("%s=a,b,c", key(defaultOpts.Type().Field(i))))
-		case reflect.Map:
-			val.Set(reflect.MakeMap(val.Type()))
-			s = append(s, fmt.Sprintf("%s=a,b,c", key(defaultOpts.Type().Field(i))))
-		default:
-			t.Errorf("Option %q has unhandled type: %v", key(defaultOpts.Type().Field(i)), val.Type())
-		}
-
-	}
-	_, _ = parseBlockOptions(strings.Join(s, " "), defaults)
-	if diff := cmp.Diff(blockOptions{}, defaults, cmp.AllowUnexported(blockOptions{}), cmpopts.EquateEmpty()); diff != "" {
-		t.Errorf("defaults appear to have been modified (-want +got):\n%s", diff)
 	}
 }
